@mmorel-35 2023-03-06
Is there any linter like `gomodguard` to ensure this rule for the whole project, if it is the intention ?