@bwplotka 2023-03-06
@mmorel-35 we usually used https://github.com/fatih/faillint for this (but probably we can't scope to just labels though)